Cleaned up streaming data functionality.
